Moved tests into one file
